# ref(profiling): improve text renderer performance