Since this is the most expensive step in our render process, squeezing out a bit of extra performance here (noticed this was starting to be a bit janky). Main change is to drop usage of the Rect class and just compute the final rect manually + a tad smarter draw heuristic

Full profile draw from devtools profiler for `static/app/utils/profiling/profile/formats/typescript/trace.json`.

Before:
<img width="516" alt="CleanShot 2022-04-29 at 08 59 50@2x" src="https://user-images.githubusercontent.com/9317857/165949034-b71dee29-cfa1-4cad-9824-ac03b3ef1916.png">

After:
<img width="551" alt="CleanShot 2022-04-28 at 19 01 26@2x" src="https://user-images.githubusercontent.com/9317857/165948739-7ab2fd91-c6b8-49ae-b7c9-1a0645f02883.png">


**Before:**
typescript (center half) x 5,175 ops/sec ±2.53% (92 runs sampled)
typescript (full profile) x 1,153 ops/sec ±2.04% (95 runs sampled)
typescript (right quarter) x 7,400 ops/sec ±2.21% (92 runs sampled)

**After:**
typescript (center half) x 12,634 ops/sec ±1.73% (93 runs sampled)
typescript (right quarter) x 14,950 ops/sec ±1.97% (93 runs sampled)
typescript (full profile) x 6,911 ops/sec ±1.62% (96 runs sampled)